Make the VoiceDialer handle "open" commands again
Bug: 2515380
The problem here was that the VoiceDialer app was assuming
that the packageName of a component can be derived by
dropping the last token of the className.  Apparently this
is not true, the packageName as far as the packageManager
is not the name of the package in the java sense.
Now when adding all of the "open" entries to the grammar,
the CommandRecognizerEngine adds both the package name
and the class name, separated by "/".

Change-Id: I79fe7d12f8f3b1b6873fcf1161b3d06f3e5e17c8