Make the voiceDialer work again
Bug: 2537307
The recognizer has a limit on the size of each semantic value
in the grammar.  Now that we are storing both the package name
and class name there, it's pretty easy to overflow that limit,
causing the grammar initialization to fail.  With this change,
only the spoken word is stored in the grammar (i.e. "calendar").
The mapping of words to package name/class name is stored
explicitly in a hashTable inside the commandRecognizerClient,
which circumvents the recognizer's fixed limit.

Bug: 2497802
If the orientation changes while an alert dialog is up,
the dialog would leak and causes an assertion failure.
With this change the VoiceDialerActivity uses the more
modern system of using dialogs, which automatically
brings down and recreates the dialog upon orientation
changes.  The BluetoothVoiceDialerActivity can't handle
this right now, it's state machine is much more
complicated.  For now, it just forces itself to be in
protrait mode all of the time.

Change-Id: I127c860b6db51426a93daf1df2d71c1c32673de5