Minor improvements to the TTS and VoiceRecognition
Now there is a short delay between when
the bluetooth SCO connection is made, and the
"speak now" utterance begins.  Prior to this,
the "speak now" was often hard to understand
because some of the audio would be cut off.

Also, the recognition does not begin
until the "speak now" utternace completes.
This should remove some of the potential
interference from the recognition.

Also, remove some code that was always setting
the TTS language to english.  Given that we've
translated the TTS strings into multiple languages
it doesn't make sense to force the pronounciation
to be in English.  This was only done because I
had copied and pasted the code from an example
and got too much of it.

Change-Id: Ib836f5041560cab2332c8392dce203ce8708be69