Fix issue 2546651: Speech indistinct/ jarred when using voice dialer over BT.
Use stream STREAM_BLUETOOTH_SCO instead of STREAM_VOICE_CALL for TTS playback.
It is more appropriate as it corresponds to current user setting for the bluetooth
headset now that STREAM_BLUETOOTH_SCO and STREAM_VOICE_CALL are not synchronized
any more.
Also limit volume to an absolute max of -18dB instead of dividing it by 2 in order
to avoid saturation without impacting lower volumes.

Change-Id: I4576de363254b666bc3d4782693592b438f1d9ed