BluetoothVoiceDialer now waits for the choices utterance to complete before listening
When the command choice returns multiple choices, the BluetoothVoiceDialer
has to speak the choices so that the user can choose which one.  Prior
to this change, it would begin listening for the user's choice at the
same time it began speaking the choices.  The recognition engine
would frequently match something before the user had said anything.
This in part because the outgoing speech can interfere with the
incoming audio.  Also, there is simply too much time during which
any random external sound can be picked up prior to the user
actually speaking.
So now, the BluetoothVoiceDialer does not begin listening for the
choice until after the choices utterance completes.

Also, improve the state handling.  In any situation where
the TTS system is speaking, the BTVoiceDialerActivity is
waiting for the utterance to complete before transitioning
to the next state and starting the recognizer.  We're not
certain that the TTS engine can be guaranteed to deliver the
utterance complete notification, so we also post a delayed runnable
to move along after a few seconds if the utterance completion
doesn't arrive.
This was not being done in a very robust way, each fallback runnable
was implemented slightly differently, and there was nothing that
would remove them if the completion did arrive.  With this change,
there is a single implementation of FallbackRunnable, which
runs exactly the same code that would have occurred on utterance completion.
Also, when the utterance really does complete, the FallbackRunnable
is removed so there should not be any way to get two completions.

Change-Id: I510be5ca65790a7c7daf53b46c56ee7feddbb8d7