Fix a problem with voicedialing on brand new phones
The code in VoiceContact runs a query to get all phones
from the contacts provider.  It expects these phones
to be grouped by the person the phone is associated with,
but it actually sorts by LastContactedTime.
The comment indicates that this is done so that in case
there are too many contacts to fit into the grammar, at
least the most recently contacted will be the first ones
entered.  These are probably the most likely ones to be
contacted again.

Normally, this works correctly.  LastContactedTime is
actually a column on the Person table, so all phones
for a single person would have the same LastContactedTime,
and ties in this value are unlikely.  However, when one gets
a new phone, or wipes data from an existing one, and
then imports contacts, all of them will have a LastContactedTime
of zero.  Since they all tie, the order they come back from in
the query is totally undefined, and can't be expected to
group multiple phones for the same person together.

This breaks an assumption, and results in multiple grammar
entries for the same person, each of which contains a single
phone number, rather than the intention, which is one row
per person, with all valid phone numbers for that person.

So on a new phone, sometimes saying something like
"call john smith" will result in only a single method
of calling that person, even if john smith has multiple
phones.

Now, the sorting is done on both LastContactedTime and
DisplayName, so in case of ties in LastContactedTime,
same names should still be grouped together.

Also, make VoiceDialerActivity and BluetoothVoiceDialerActivity
use a new API to acquire and abandon Audio Focus.  This
is an improved way of turning other media playback while
voice dialing, and allowing it to turn back on when finished.

Also correct some minor spelling errors.

Change-Id: Icc71bcf104bf2c02fdd8c03aa2749640561248e7