Make the BluetoothVoiceDialerActivity and VoiceDialerActivity into one activity
These had been separated into two activites while the BluetoothVoiceDialer
was being worked on, but it's a lot of duplicate code.  Now there is a single
activity that can run in either bluetooth or non-bluetooth mode, depending
on how it is started.
The state transitions are slightly simpler when running in non-bluetooth
mode, because it does not use text to speech to inform the user about
what it is doing.

Change-Id: Iddf84108a84420917073fa18999d7394ca308400