Improved fix for Bug: 2326485
When starting the voiceDialer, there is a tone played
to indicate that the microphone is listening.  It's
important the tone not be picked up by the mic and
sent to the recognizer engine, or it will corrupt
the sample.  Prior to this change, the VoiceDialerActivity
would simply pull a fixed 380 milliseconds of audio
from the microphone inputStream.  This works, but
may not be reliable on all hardware platforms because
the size of audio buffers and performance of the audio
system may vary.
Now, the VoiceDialerActivity removes audio samples
and attempts to detect the tone.  When the tone ends,
it stops pulling samples.  If no tone is detected within
500 milliseconds, it times out and stops removing samples.

Change-Id: Ic57ba1947fa0d315bec081bf03a32abb1a822fdf