Fix Bug: 2606879
The problem is that if the user leaves the VoiceDialerActivity
using the home button, when it is started again it never starts
listening.  This is because all of the initialization is done
in onCreate, but the teardown is done in onStop.  When the
activity is left using the home button, the activity is stopped,
but not destroyed, so onCreate is never called when the activity
is started up again.

This change moves the initialization to onStart, and leaves teardown
in onStop. It really doesn't make any sense for this particular activity
to be stopped or paused, if it's stopped at all it should be finished.
So now the onStop function also calls finish().

Change-Id: I4ddc7e329dfbf88d7b732191581877d9e5a3f0a3