Fix Bugs 2626599 and 2615852
Prior to this change, VoiceDialerActivity would decide
whether to run in bluetooth mode based upon the intent
that started it.  If it was ACTION_VOICE_COMMAND, then
it would run in bluetooth mode.
There are two problems with this:
1.  If there is a bluetooth headset connected, but
the voicedialer is launched from the menu, then
it will come up in non-bluetooth mode, so the audio
would not be routed correctly and it would not prompt
the user via TTS.
2.  If there is not a bluetooth headset connected, but
and then something other than the headset fires the
ACTION_VOICE_COMMAND, it will come up in bluetooth mode,
and then block indefintely trying to connect to the
headset.

With this change, the VoiceDialerActivity always creates
a BluetoothHeadset object.  It waits for the first onServiceConnected
callback to decide which mode to run in.  Now it runs in the
correct mode, regardless of how it was launched.

Change-Id: Ic8e37ee094d6bd84ec8171e6d64d5cfda1a33b29