Recognize phone numbers without supported types
Currently VoiceDialer cannot show a candidate when it has a type
the recognition engine cannot recognize. For example if a
contact has one phone number with the type CUSTOM, VoiceDialer
never shows the person as a candidate even when a user ask so
("Call XXX").

This change prepares "fallback" phone number id and use it when
there's no other available ids with recognizable types, to support
the case above.

TESTED:
- Have a contact (say "Ada" here) with several phone numbers one
  of which must have a custom type which isn't supported by
  VoiceDialer ("CUSTOM"). Say "Call Ada" and see if that custom
  phone number appears as another entry ("Call Ada")
- Try "Call Ada at home" or other voice command and see they work
  as before.

Bug: 5585945
Change-Id: Idd0e539209cd14e786007e8ce0b121c165994c5a